fix: allow `MaxWidth` enum instance
The `getMaxWidth()` method is compatible with handling the `MaxWidth` enum. Previously, it was not possible to use `MaxWidth` enum instances when overwriting the `$maxWidth` property.

Note: This change was not applied to the `3.x` branch as it introduces a breaking change. The change may impact classes extending `SimplePage` or `EditProfile` if the `maxWidth` property has already been overridden.